Update docker images so they build
I used a newer ubuntu base image and swapped docker for podman. I also
replaced the github link with my own repo. And I removed the noreturn
attribute on the cpu::exception function because this was causing
problems for the aspectc compiler, which I also updated to the recent
2.3 version. The base and generic-tracing docker images should build out
of the box now.